Add deleteImagesTrigger when tags are deleted. Also add type annotation to existing tags and rename options fields directory to prefix
